+++ datadog_checks_dev/datadog_checks/dev/tooling/constants.py
@@ -83,10 +83,6 @@ NOT_TILES = [
     'wmi_check',
 ]
 
-# If a file changes in a PR with any of these file extensions,
-# a test will run against the check containing the file
-TESTABLE_FILE_PATTERNS = ('*.py', '*.ini', '*.in', '*.txt', '*.yml', '*.yaml', '**/tests/*')
-NON_TESTABLE_FILES = ('auto_conf.yaml', 'agent_requirements.in')
 
 REQUIREMENTS_IN = 'requirements.in'
